+++ MercadoPagoSDK/MercadoPagoSDK/Flows/OneTap/UI/Components/PXSummaryComposer+AddRow.swift
@@ -49,7 +49,7 @@ extension PXSummaryComposer {
     }
 
     func totalToPayRow() -> PXOneTapSummaryRowData {
-        let totalAmountToShow = Utils.getAmountFormated(amount: amountHelper.getAmountToPayWithoutPayerCost(paymentOptionID: selectedCard?.cardId, paymentMethodId: selectedCard?.selectedApplication?.paymentMethodId, paymentTypeId: selectedCard?.selectedApplication?.paymentTypeId), forCurrency: currency)
+        let totalAmountToShow = Utils.getAmountFormated(amount: amountHelper.amountToPay, forCurrency: currency)
         let row = PXOneTapSummaryRowData(
             title: "total_row_title_default".localized,
             value: totalAmountToShow,